FIX : JWT 추가에 따른 Controller 수정

+++ src/main/java/Study/ApiStudy/controller/BoardController.java
@@ -1,10 +1,12 @@
 package Study.ApiStudy.controller;
 
+import Study.ApiStudy.config.auth.PrincipalDetails;
 import Study.ApiStudy.dto.BoardDto;
 import Study.ApiStudy.entity.User;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import Study.ApiStudy.repository.UserRepository;
 import Study.ApiStudy.response.Response;
@@ -36,8 +38,9 @@ public class BoardController {
     @ApiOperation(value = "게시글 작성", notes = "게시글을 작성한다.")
     @ResponseStatus(HttpStatus.OK)
     @PostMapping("/board/write")
-    public Response write(@RequestBody BoardDto boardDto){
-        User user = userRepository.findById(1).get();
+    public Response write(@RequestBody BoardDto boardDto, Authentication authentication){
+        PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
+        User user = principalDetails.getUser();
         return new Response("성공","글 작성 성공",boardService.write(boardDto,user));
     }
 
@@ -45,17 +48,25 @@ public class BoardController {
     @ApiOperation(value = "게시글 수정", notes = "게시글을 수정한다.")
     @ResponseStatus(HttpStatus.OK)
     @PutMapping("/board/update/{id}")
-    public Response edit(@RequestBody BoardDto boardDto, @PathVariable("id") Integer id){
-        User user = userRepository.findById(1).get(); // 임시로
-        return new Response("성공","글 수정 성공",boardService.update(id,boardDto));
+    public Response edit(@RequestBody BoardDto boardDto, @PathVariable("id") Integer id, Authentication authentication){
+        PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
+        User user = principalDetails.getUser();
+        if(user.getName().equals(boardService.getBoard(id).getWriter())) return new Response("성공","글 수정 성공",boardService.update(id,boardDto));
+        else return new Response("실패", "본인 게시물만 수정 할 수 있습니다.",null);
     }
 
     // 게시글 삭제
     @ApiOperation(value = "게시글 삭제", notes = "게시글을 삭제한다")
     @ResponseStatus(HttpStatus.OK)
     @DeleteMapping("/board/delete/{id}")
-    public Response delete(@PathVariable("id") Integer id){
-        boardService.delete(id);
-        return new Response("성공","글 삭제 성공",null);
+    public Response delete(@PathVariable("id") Integer id,Authentication authentication){
+        PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
+        User user = principalDetails.getUser();
+        if(user.getName().equals(boardService.getBoard(id).getWriter())){
+            boardService.delete(id);
+            return new Response("성공","글 삭제 성공",null);
+        }else{
+            return new Response("실패","본인 게시글만 삭제할 수 있습니다.",null);
+        }
     }
 }
