+++ src/main/java/Study/ApiStudy/controller/CommentController.java
@@ -1,5 +1,6 @@
 package Study.ApiStudy.controller;
 
+import Study.ApiStudy.config.auth.PrincipalDetails;
 import Study.ApiStudy.dto.CommentDto;
 import Study.ApiStudy.entity.User;
 import Study.ApiStudy.repository.UserRepository;
@@ -8,6 +9,7 @@ import Study.ApiStudy.service.CommentService;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 @RequiredArgsConstructor
@@ -20,9 +22,9 @@ public class CommentController {
     @ApiOperation(value = "댓글 작성", notes = "댓글을 작성한다.")
     @ResponseStatus(HttpStatus.OK)
     @PostMapping("/comments/{boardId}")
-    public Response writeComment(@PathVariable("boardId") Integer boardId, @RequestBody CommentDto commentDto){
-        User user = userRepository.findById(1).get();
-        return new Response("성공","댓글 작성을 완료했습니다.",commentService.writeComment(boardId,commentDto,user));
+    public Response writeComment(@PathVariable("boardId") Integer boardId, @RequestBody CommentDto commentDto, Authentication authentication){
+        PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
+        return new Response("성공","댓글 작성을 완료했습니다.",commentService.writeComment(boardId,commentDto, principalDetails.getUser()));
     }
 
     // 게시글에 달린 댓글 모두 불러오기
@@ -37,9 +39,13 @@ public class CommentController {
     @ApiOperation(value = "댓글 삭제하기", notes = "게시글에 달린 댓글을 삭제합니다.")
     @ResponseStatus(HttpStatus.OK)
     @DeleteMapping("/comments/{boardId}/{commentId}")
-    public Response deleteComment(@PathVariable("boardId") Integer boardId,@PathVariable("commentId") Integer commentId){
-        return new Response("성공","댓글 삭제 완료",commentService.deleteComment(commentId));
+    public Response deleteComment(@PathVariable("boardId") Integer boardId,@PathVariable("commentId") Integer commentId, Authentication authentication){
+        PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
+        if(commentService.getComment(commentId).getWriter().equals(principalDetails.getUser().getName())) return new Response("성공","댓글 삭제 완료",commentService.deleteComment(commentId));
+
+        return new Response("실패","댓글 작성자가 아닙니다.",null);
     }
+
 
 
 }
